# openapi.yaml
openapi: 3.0.3
info:
  title: Face Sequencer Pro API
  version: c6f429b
  description: Modular OpenAPI specification (assembled)
servers:
  - url: http://localhost:5000
paths:
...
  /api/sequence/frame/{frame_id}:
    get:
      tags:
        - sequence
      summary: Retrieve a single sequence frame preview
      parameters:
        - in: path
          name: frame_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Frame data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/StandardResponse"
                  - type: object
                    properties:
                      char:
                        type: string
                      duration:
                        type: integer
                      is_pause:
                        type: boolean
                      image:
                        type: string
                      dimensions:
                        type: array
                        items:
                          type: integer
        "404":
          description: Frame not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sequence/update:
    post:
      tags:
        - sequence
      summary: Update sequence frame properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SequenceUpdateRequest"
      responses:
        "200":
          description: Frame updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
        "400":
          description: Invalid update payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Frame not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sequence/reorder:
    post:
      tags:
        - sequence
      summary: Reorder sequence frames
      requestBody:
        required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SequenceReorderRequest"
      responses:
        "200":
          description: Sequence reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
        "400":
          description: Invalid reorder payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sequence/delete:
    post:
      tags:
        - sequence
      summary: Delete a sequence frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SequenceDeleteRequest"
      responses:
        "200":
          description: Frame deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
        "400":
          description: Invalid delete payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Frame not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: string
        error_type:
          type: string
        details:
          type: object
        recommendations:
          type: array
          items:
            type: string
        lifecycle_stage:
          type: string
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          required:
            - success
            - error
          properties:
            success:
              const: false
            error:
              type: string
            error_type:
              type: string
            details:
              type: object
            recommendations:
              type: array
              items:
                type: string
            lifecycle_stage:
              type: string
...
    SequenceUpdateRequest:
      type: object
      required:
        - frame_id
        - updates
      properties:
        frame_id:
          type: integer
          minimum: 0
        updates:
          type: object
          properties:
            duration:
              type: integer
              minimum: 1
    SequenceReorderRequest:
      type: object
      required:
        - from_index
        - to_index
      properties:
        from_index:
          type: integer
          minimum: 0
        to_index:
          type: integer
          minimum: 0
    SequenceDeleteRequest:
      type: object
      required:
        - frame_id
      properties:
        frame_id:
          type: integer
          minimum: 0
...